Fix tests about iterables
